aaaaaaaaaaaaaaaaaaa help

+++ src/main/java/org/generation/italy/model/dto/RegistrationRequestEvaluationDto.java
@@ -2,6 +2,6 @@ package org.generation.italy.model.dto;
 //Lo vede l'utente non l'admin
 //Serve per i dati che l'utente manda per modificare la form
 public record RegistrationRequestEvaluationDto(
-        Long sessionId,
-        Long sessionTypeId
+        Integer sessionId,
+        Integer sessionTypeId
 ) {}
